meditrack_project/requirements.txt
# Core Django
Django==4.2.7

# Django REST Framework
djangorestframework==3.14.0


# Filtering, Search, Ordering (P10)
django-filter==23.5

# Swagger / OpenAPI (P12)
drf-spectacular==0.27.1

# CORS untuk frontend terpisah (Decoupled)
django-cors-headers==4.3.1

# Utility
pytz==2023.3.post1
sqlparse==0.4.4
asgiref==3.7.2

# Env config (opsional tapi rapi)
python-dotenv==1.0.0
